refactor(Piece): piece 패키지 분리
